flatten the structure, get rid of the separate commitments package
