--- castai/commitments/fields.go
@@ -1,10 +1,0 @@
-package commitments
-
-const (
-	FieldAzureReservationsCSV = "azure_reservations_csv"
-	FieldGCPCUDsJSON          = "gcp_cuds_json"
-
-	FieldAzureReservations = "azure_reservations"
-	FieldGCPCUDs           = "gcp_cuds"
-	FieldGCPCUDConfigs     = "gcp_cud_configs"
-)
