+++ castai/resource_commitments_mapping_test.go
@@ -1,4 +1,4 @@
-package commitments
+package castai
 
 import (
 	"errors"
